+++ src/main/java/net/soulsweaponry/networking/C2S/receivers/KeybindAbilityC2SReceiver.java
@@ -43,13 +43,13 @@ public class KeybindAbilityC2SReceiver {
             //TODO remove above
             for (ItemStack armorStack : player.getArmorItems()) {
                 if (armorStack.getItem() instanceof IHasAbilities abilityItem) {
-                    abilityItem.useKeybindAbilityServer(serverWorld, armorStack, player);
+                    abilityItem.useKeybindAbilityServer(serverWorld, armorStack, player, null);
                 }
             }
             for (Hand hand : Hand.values()) {
                 ItemStack stack = player.getStackInHand(hand);
                 if (stack.getItem() instanceof IHasAbilities abilityItem) {
-                    abilityItem.useKeybindAbilityServer(serverWorld, stack, player);
+                    abilityItem.useKeybindAbilityServer(serverWorld, stack, player, hand);
                 }
             }
         });